[web] Change file names to cebab lower case. Delete pretier from package.json
