+++ package.json
@@ -19,7 +19,6 @@
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
     "@vitejs/plugin-react": "^3.0.0",
-    "prettier": "2.8.2",
     "typescript": "^4.9.3",
     "vite": "^4.0.0"
   }
