+++ src/components/Button/Button.tsx
@@ -1,6 +1,6 @@
 import cn from 'classnames';
 
-import styles from './Button.module.css';
+import styles from './button.module.css';
 
 interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   priority: string;
