+++ src/components/Button/index.ts
@@ -1,1 +1,1 @@
-export { Button } from './Button';
+export { Button } from './button';
