+++ src/components/Input/index.ts
@@ -1,1 +1,1 @@
-export { Input } from './Input';
+export { Input } from './input';
