+++ src/pages/Home/index.ts
@@ -1,1 +1,1 @@
-export { Home } from './Home';
+export { Home } from './home';
