--- src/pages/SignPages/index.ts
@@ -1,2 +1,0 @@
-export { SignIn } from './SignIn';
-export { SignUp } from './SignUp';
