+++ src/pages/sign-pages/index.ts
@@ -1,0 +1,2 @@
+export { SignIn } from './sign-in';
+export { SignUp } from './sign-up';
