+++ src/pages/sign-pages/sign-in.tsx
@@ -1,21 +1,21 @@
 import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { Button } from '../../components/Button';
-import { Input } from '../../components/Input';
+import { Button } from '../../components/button';
+import { Input } from '../../components/input';
 
-import styles from './Sign.module.css';
+import styles from './sign.module.css';
 
-export function SignUp() {
+export function SignIn() {
   const navigate = useNavigate();
 
   useEffect(() => {
     if(document.cookie.split(';').some((item) => item.trim().startsWith('signedin='))) {
        navigate('/');
     };
-  },[])
+  }, []);
 
-  function login():void {
+  function login() {
     document.cookie = 'signedin=1';
     navigate('/');
   }
@@ -28,7 +28,7 @@ export function SignUp() {
           onSubmit={login}
         >
           <h2 className={styles.form_title}>
-            Sign Up
+            Sign In
           </h2>
           <Input
             type='text'
@@ -42,18 +42,6 @@ export function SignUp() {
           >
             Password
           </Input>
-          <Input
-            type='password'
-            required
-          >
-            Confirm Password
-          </Input>
-          <Input
-            type='email'
-            required
-          >
-            Mail
-          </Input>
           <p className={styles.form_error}>
             Error example
           </p>
@@ -61,13 +49,13 @@ export function SignUp() {
             priority='primary'
             type='submit'
           >
-            Sign Up
+            Sign In
           </Button>
           <a 
-            href="/signin"
+            href="/signup"
             className={styles.form_link}
           >
-              Already have an accaunt
+              Create new accaunt
           </a>
         </form>
       </div>
